[TACHYON-1673] Make chgrpr, chownr, chmodr as cmd alias and remove unused classes.
