+++ shell/src/main/java/tachyon/shell/command/ChownCommand.java
@@ -25,13 +25,15 @@ import org.apache.commons.cli.Options;
 
 import tachyon.TachyonURI;
 import tachyon.client.file.FileSystem;
+import tachyon.client.file.options.SetAclOptions;
 import tachyon.conf.TachyonConf;
+import tachyon.exception.TachyonException;
 
 /**
  * Changes the owner of a file or directory specified by args.
  */
 @ThreadSafe
-public final class ChownCommand extends AbstractAclCommand {
+public final class ChownCommand extends AbstractTfsShellCommand {
 
   /**
    * Creates a new instance of {@link ChownCommand}.
@@ -65,6 +67,25 @@ public final class ChownCommand extends AbstractAclCommand {
 
     opts.addOption(recursive);
     return opts;
+  }
+
+  /**
+   * Changes the owner for the directory or file with the path specified in args.
+   *
+   * @param path The {@link TachyonURI} path as the input of the command
+   * @param owner The owner to be updated to the file or directory
+   * @param recursive Whether change the owner recursively
+   * @throws IOException if command failed
+   */
+  protected void chown(TachyonURI path, String owner, boolean recursive) throws IOException {
+    try {
+      SetAclOptions options = SetAclOptions.defaults().setOwner(owner).setRecursive(recursive);
+      mFileSystem.setAcl(path, options);
+      System.out.println("Changed owner of " + path + " to " + owner);
+    } catch (TachyonException e) {
+      throw new IOException("Failed to changed owner of " + path + " to " + owner + " : "
+          + e.getMessage());
+    }
   }
 
   @Override
